Signature hash generation (#273)
* - fix signature hash generation;

* - update bloom filter;

* - fix test;

+++ revocation/src/main/java/dcc/app/revocation/validation/BloomFilter.java
@@ -10,8 +10,9 @@ import java.io.OutputStream;
 
 import dcc.app.revocation.validation.exception.FilterException;
 
+
 public interface BloomFilter {
-    float getP();
+    double getP();
 
     int getK();
